Don't assume each reference contains one record.
A progressbar was assuming that the number of
reference files provided was equal to the number of
fasta records in the validated_ref file. In fact, each
fasta reference file could contain more than one record.
Providing a reference with more than one record in it
was causing a traceback when the progress bar was
updated to a number higher than its maximum value.

This patch changes the progress bar to be based on the
number of records in the validated_ref file. This means
using a list rather than an iterator, so might cause an
increase in memory usage, but I think this will be very
small and temporary, and a similar list is used
earlier in processRefFasta.