Make sure file is closed on exception.
Previously, this piece of code wasn't closing files when
there was an exception. The traditional way to do this was
with a 'finally' clause, but these days opening the file
with 'with' is the standard way.

I was seeing a strange issue here where I would be unable
to import Pool from the multiprocessing module in the standard
library. What was in fact happening was that we'd run out of
file handles, but this was being hidden by the python's import
mechanism.